Fix Source files for a project sometimes open in the miscellaneous project
Fixes #489

Due to CPS bug (#274485) - we were getting incorrect capabilities set when we got called back for the first evaluation, resulting in our "handlers" being empty. Worked around this by setting "CSharp" and "LanguageService" as part of the default set of capabilities for csproj/vbproj (not shproj). We can remove this when CPS bug is fixed.
